feat(builder): Move dtfs building to a separate file
